+++ src/primitives/test/primitives.test.js
@@ -4,7 +4,7 @@ import 'react-primitives';
 import { Text, View } from 'react-primitives';
 import TestRenderer from 'react-test-renderer';
 
-import styled from '../index';
+import styled, { ThemeProvider } from '../index';
 
 // for the purpose of testing warnings we want to make sure they're always fired
 jest.mock('../../utils/once', () => cb => cb);
@@ -245,6 +245,25 @@ describe('primitives', () => {
 
       expect(text.props).toMatchObject({
         children: 'Something else',
+        style: [{}],
+      });
+    });
+
+    it('function form allows access to theme', () => {
+      const Comp = styled.Text.attrs(props => ({
+        'data-color': props.theme.color,
+      }))``;
+
+      const wrapper = TestRenderer.create(
+        <ThemeProvider theme={{ color: 'red' }}>
+          <Comp>Something else</Comp>
+        </ThemeProvider>
+      );
+      const text = wrapper.root.findByType('Text');
+
+      expect(text.props).toMatchObject({
+        children: 'Something else',
+        'data-color': 'red',
         style: [{}],
       });
     });